Norm'd the file layout sturcture & added inline_video prototypes
